+++ modules/tinymce/src/core/main/ts/annotate/AnnotationChanges.ts
@@ -6,10 +6,12 @@
  */
 
 import { Arr, Cell, Obj, Optional, Optionals, Singleton, Throttler } from '@ephox/katamari';
+import { Attribute } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
 import { AnnotationsRegistry } from './AnnotationsRegistry';
-import { identify } from './Identification';
+import { findMarkers, identify } from './Identification';
+import { dataAnnotationActive } from './Markings';
 
 export interface AnnotationChanges {
   readonly addListener: (name: string, f: AnnotationListener) => void;
@@ -24,7 +26,7 @@ export interface AnnotationListenerData {
 
 export type AnnotationListenerMap = Record<string, AnnotationListenerData>;
 
-const setup = (editor: Editor, _registry: AnnotationsRegistry): AnnotationChanges => {
+const setup = (editor: Editor, registry: AnnotationsRegistry): AnnotationChanges => {
   const changeCallbacks = Cell<AnnotationListenerMap>({ });
 
   const initData = (): AnnotationListenerData => ({
@@ -62,26 +64,38 @@ const setup = (editor: Editor, _registry: AnnotationsRegistry): AnnotationChange
     });
   };
 
+  const toggleActiveAttr = (uid: string, state: boolean) => {
+    Arr.each(findMarkers(editor, uid), (span) => {
+      if (state) {
+        Attribute.set(span, dataAnnotationActive(), 'true');
+      } else {
+        Attribute.remove(span, dataAnnotationActive());
+      }
+    });
+  };
+
   // NOTE: Runs in alphabetical order.
   const onNodeChange = Throttler.last(() => {
-    const callbackMap = changeCallbacks.get();
-    const annotations = Arr.sort(Obj.keys(callbackMap));
+    const annotations = Arr.sort(registry.getNames());
     Arr.each(annotations, (name) => {
       updateCallbacks(name, (data) => {
         const prev = data.previous.get();
         identify(editor, Optional.some(name)).fold(
           () => {
-            if (prev.isSome()) {
+            prev.each((uid) => {
               // Changed from something to nothing.
               fireNoAnnotation(name);
               data.previous.clear();
-            }
+              toggleActiveAttr(uid, false);
+            });
           },
           ({ uid, name, elements }) => {
             // Changed from a different annotation (or nothing)
             if (!Optionals.is(prev, uid)) {
+              prev.each((uid) => toggleActiveAttr(uid, false));
               fireCallbacks(name, uid, elements);
               data.previous.set(uid);
+              toggleActiveAttr(uid, true);
             }
           }
         );